Added verse server
The original .zip file from QuelSolaar.com did not include
a verse server implementation.

However, a .zip file originally at the url:

    http://web.archive.org/web/20120204174225/http://verse.blender.org/uploads/media/verse-r5p0.zip

does contain a server implementation. It appears this code
predates the version of the verse API on QuelSolaar.com by
approximately 1 year.

With some minor changes (one function) I was able to compile the server using
the verse api and header from the original .zip file on QuelSolaar.com.

The resulting binary "works" in that loq_airou connects to it
and seems to behave reasonably. I expect that there are security
issues and bugs which have been introduced due to the mashup of old
and new code. It remains to learn more about the verse protocol
and extend the server, possibly with a from-scratch implementation
using the existing code as a guide.