Added verse server from blender.org git repo; recreated build scripts.
The purpose of this commit is to remove the verse server
I added before (which was an older version) and use the
version from blender. The reasoning here is that I prefer
to start from a point where I know someone intended that
the server and the API should be compatible.

After this change, no code in the source/verse directory comes
from the quelsolaar.com zip file; it's all from
blender.

Additionally, I rewrote the build scripts to automatically
run the verse api generation tool as necessary. Because
of this, I deleted the files which are generated at build
time and therefore shouldn't be edited. (Edit the inputs instead).

If necessary I will re-commit these files later as documentation,
in a documentation directory. It will then be the responsibility
of the documentation generation scripts to keep them in sync
with the input files.