Working on win32 build script.
I have added a new config file tupconf/mingw32.config. This configuration
is for cross compilation of win32 binaries from an linux host. I will also be
supporting native compilation on a windows host using mingw, but this is not
yet implemented. The plan here is to use the cross-compilation target to
produce windows binaries via the CI script.

The verse-gen-tool still uses the native toolchain as we need to run it
during the build process.